requirements.txt
fastapi==0.115.9
uvicorn==0.34.2
python-multipart==0.0.20
chromadb==1.0.0
openai==1.76.2
pydantic==2.11.4
python-dotenv==1.1.0
sounddevice==0.5.1
numpy==1.26.4
scipy==1.15.2
openai-whisper==20240930
elevenlabs==2.3.0
sentence-transformers==4.1.0
chroma-haystack==3.1.0
haystack-ai==2.13.1
requests==2.32.3
textblob==0.19.0